made expo secure store adapter a reusable
